+++ src/main/java/com/onevizion/uitest/api/helper/entity/EntityConfigApp.java
@@ -271,7 +271,7 @@ public class EntityConfigApp {
         window.closeModal(By.id(AbstractSeleniumCore.BUTTON_CANCEL_ID_BASE));
     }
 
-    public void testInGrid(Long gridId, Long rowIndex, ConfigApp configApp) {
+    public void testInGrid(Long gridId, int rowIndex, ConfigApp configApp) {
         Map<Integer, String> gridVals = new HashMap<>();
 
         gridVals.put(js.getColumnIndexByLabel(gridId, "Trackor Type"), configApp.getTrackorType());
